Refactor Phase 2: The deconstruction of gfx.h
