+++ .gitmodules
@@ -1,7 +1,6 @@
+[submodule "GameSystem/external/mpu6050-api"]
+	path = GameSystem/external/mpu6050-api
+	url = https://github.com/AidanRandtoul/MPU6050-C-CPP-Library-for-Raspberry-Pi
 [submodule "GameSystem/external/matrix-api"]
 	path = GameSystem/external/matrix-api
-	url = https://github.com/hzeller/rpi-rgb-led-matrix.git
-	ignore = dirty
-[submodule "GameSystem/external/MPU6050-C-CPP-Library-for-Raspberry-Pi"]
-	path = GameSystem/external/MPU6050-C-CPP-Library-for-Raspberry-Pi
-	url = https://github.com/AidanRandtoul/MPU6050-C-CPP-Library-for-Raspberry-Pi
+	url = https://github.com/hzeller/rpi-rgb-led-matrix
