+++ GameSystem/include/loading-bar.h
@@ -1,0 +1,21 @@
+#pragma once
+#include "object.h"
+
+class LoadingBar : public Object
+{
+public:
+    float diameter;
+    int height;
+    void setDiameter(float d);
+
+public:
+    float getDiameter();
+    int getHeight();
+    LoadingBar(float x_pos, float y_pos, int h)
+        : Object{x_pos, y_pos, BAR}
+    {
+        instances.push_back(this);
+        diameter = 0;
+        height = h;
+    }
+};
