+++ GameSystem/include/render.h
@@ -1,0 +1,14 @@
+#pragma once
+#include "led-matrix.h"
+#include "graphics.h"
+
+class Marple;
+class Object;
+
+void fillRect(float start_x, float start_y, int w, int h, Object *obj, Object *(&array)[64][64]);
+
+void fillBorder(rgb_matrix::Color borderColour, int width);
+
+void ColliderCheck(Marple * marple);
+
+void clearAll(rgb_matrix::Canvas *c);
