+++ GameSystem/source/loading-bar.cpp
@@ -1,0 +1,16 @@
+#include "loading-bar.h"
+
+void LoadingBar::setDiameter(float d)
+{
+    diameter = d;
+}
+
+float LoadingBar::getDiameter()
+{
+    return diameter;
+}
+
+int LoadingBar::getHeight()
+{
+    return height;
+}
